backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
google-generativeai==0.8.5
gunicorn==22.3.1

